Support database selection
In order to allow the use of different databases in Redis, the
connection now supports switching to a new database.

One important factor to understand here is that if you are using
`pconnect` and you do a `select` operation, you may be changing the same
connection the rest of your code is using.